Implementação da caixa fantasma

+++ src/main/physics/Ray.java
@@ -1,31 +1,31 @@
 package main.physics;
 
-import javax.vecmath.Vector3d;
+import javax.vecmath.Vector3f;
 
 public class Ray {
 
-	private Vector3d origin;
-	private Vector3d direction;
+	private Vector3f origin;
+	private Vector3f direction;
 	
-	public Ray(Vector3d origin, Vector3d direction) {
+	public Ray(Vector3f origin, Vector3f direction) {
 		super();
 		this.origin = origin;
 		this.direction = direction;
 	}
 	
-	public Vector3d getOrigin() {
+	public Vector3f getOrigin() {
 		return origin;
 	}
 	
-	public void setOrigin(Vector3d origin) {
+	public void setOrigin(Vector3f origin) {
 		this.origin = origin;
 	}
 	
-	public Vector3d getDirection() {
+	public Vector3f getDirection() {
 		return direction;
 	}
 	
-	public void setDirection(Vector3d direction) {
+	public void setDirection(Vector3f direction) {
 		this.direction = direction;
 	}
